ping: Fix corruption of hostname string with optimizations enabled
The problem was in assigning a pointer to 'hnamebuf[]' defined in
scope of an 'else' branch block to 'vars.hostname' which had a larger
scope. The compiler reused that memory location for the 'types[]'
array of pointers in another block and the string 'vars.hostname' got
corrupted when 'types[]' was written to.