ping: tests: Fix use of memory for test_argv[] array
The body of ARGV_SET_FROM_EXPR() and ARGV_SET_LDBL_FROM_EXPR() macros
cannot be in a block because pointer assigned to test_argv[i] points
to memory local for this block. The memory can be reused after exiting
the block and rewrited. When the body is used without a block,
variable naming conflicts needs to resolved if the same macro is used
more than once. This is solved by appending an index to each variable
in the macros.